+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Menu drawer not working well on small screens
 
+### Added
+- Optimistic rating update
+
 ## [0.2.0] - 2025-02-21
 
 ### Added